requirements.txt
flask
requests
PyMuPDF
pinecone
sentence-transformers
python-dotenv
